Working on RecipeForm and recipeEdit

+++ src/recipe/ShowRecipe.js
@@ -23,12 +23,12 @@ const ShowRecipe = ({ recipe, onDelete }) => {
 
     const deleteRecipe = async () => {
         const user = localStorage.getItem("user");
-        // if (user) {
-            // const u = JSON.parse(user);
+        if (user) {
+            const u = JSON.parse(user);
             let response = await fetch(`http://localhost:8080/project/recipe/${recipe.id}`, {
                 method: "DELETE",
                 headers: {
-                    // Authorization: u.token,
+                    Authorization: u.token,
                     "Accept": "application/json",
                     "Content-Type": "application/json",
                 }
@@ -40,7 +40,7 @@ const ShowRecipe = ({ recipe, onDelete }) => {
             } else {
                 console.log("Error while deleting recipe");
             }
-        // }
+        }
     }
 
     return (
@@ -99,15 +99,15 @@ const ShowRecipe = ({ recipe, onDelete }) => {
 
                 <Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
                     {/* {user && user.role === "ROLE_REGULAR_USER" || user.role === "ROLE_CHEF" || user.role === "ROLE_ADMIN" ? */}
-                        <Tooltip title="Info">
-                            <IconButton
-                                sx={{ margin: '0px 8px 15px 8px', color: '#6bb187', fontSize: '1em' }}
-                                aria-label="info"
-                                onClick={() => navigate(`recipe_details/${recipe.id}`)}>
-                                <InfoIcon />
-                            </IconButton>
-                        </Tooltip> 
-                        {/* :<>
+                    <Tooltip title="Info">
+                        <IconButton
+                            sx={{ margin: '0px 8px 15px 8px', color: '#6bb187', fontSize: '1em' }}
+                            aria-label="info"
+                            onClick={() => navigate(`recipe_details/${recipe.id}`)}>
+                            <InfoIcon />
+                        </IconButton>
+                    </Tooltip>
+                    {/* :<>
                             <Tooltip title="Info">
                                 <IconButton
                                     sx={{ margin: '0px 8px 15px 8px', color: '#6bb187' }}
@@ -117,8 +117,8 @@ const ShowRecipe = ({ recipe, onDelete }) => {
                                 </IconButton>
                             </Tooltip>
                         </> */}
-                         {/* } */}
-                    {/* {user && user.role === "ROLE_ADMIN" || user.role === "ROLE_CHEF" ? */}
+                    {/* } */}
+                    {user && user.role === "ROLE_ADMIN" || user.role === "ROLE_CHEF" ?
                         <>
                             <Tooltip title="Edit">
                                 <IconButton sx={{ margin: '0px 8px 15px 8px', color: '#6bb187' }} aria-label="edit" onClick={() => navigate(`edit_recipe/${recipe.ID}`)}>
@@ -129,16 +129,16 @@ const ShowRecipe = ({ recipe, onDelete }) => {
                                 <IconButton sx={{ margin: '0px 8px 15px 8px', color: '#6bb187' }} aria-label="delete" onClick={deleteRecipe}>
                                     <DeleteIcon />
                                 </IconButton>
-                            </Tooltip> </> 
-                            {/* : <></>} */}
+                            </Tooltip> </>
+                        : <></>}
                     {/* {user && user.role === "ROLE_REGULAR_USER" ? */}
-                        <>
-                            <Tooltip title="Add to my Cookbook">
-                                <IconButton sx={{ margin: '0px 8px 15px 8px', color: '#6bb187' }} aria-label="Add to my Cookbook">
-                                    <FavoriteBorderIcon />
-                                </IconButton>
-                            </Tooltip> </> 
-                            {/* : <></>} */}
+                    <>
+                        <Tooltip title="Add to my Cookbook">
+                            <IconButton sx={{ margin: '0px 8px 15px 8px', color: '#6bb187' }} aria-label="Add to my Cookbook">
+                                <FavoriteBorderIcon />
+                            </IconButton>
+                        </Tooltip> </>
+                    {/* : <></>} */}
                 </ Box>
             </Card>
         </Grid>
